pyproject.toml
[tool.poetry]
name = "pytest-shared-session-scope"
version = "0.1.1"
description = ""
authors = ["Stefan Rasmussen <stefan@bruhn.io>"]
readme = "README.md"
packages = [{include = "pytest_shared_session_scope", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10"
pytest = ">=7"
filelock = ">=2"
typing-extensions = ">=3.6.2"

[tool.poetry.plugins.pytest11]
pytest-shared-session-scope = "pytest_shared_session_scope.fixtures"


[tool.poetry.group.dev.dependencies]
ruff = "0.5.6"
pytest-cov = "5.0.0"
pytest-xdist = "3.6.1"

# pyproject.toml
[tool.pytest.ini_options]
addopts = "-n auto"
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
